requirements/base.txt
# SHA1:c33b9624e39da6ea71996ef770d2d17d12282f19
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
attrs==22.2.0
    # via jsonschema
certifi==2022.12.7
    # via requests
chardet==5.1.0
    # via trimesh
charset-normalizer==3.0.1
    # via requests
cloudpickle==2.2.1
    # via gym
colorlog==6.7.0
    # via trimesh
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
cython==0.29.33
    # via python-fcl
fonttools==4.38.0
    # via matplotlib
gym==0.23.1
    # via -r requirements/base.in
gym-notices==0.0.8
    # via gym
idna==3.4
    # via requests
importlib-metadata==6.0.0
    # via gym
importlib-resources==5.10.2
    # via jsonschema
jsonschema==4.17.3
    # via trimesh
kiwisolver==1.4.4
    # via matplotlib
lxml==4.9.2
    # via trimesh
mapbox-earcut==1.0.1
    # via trimesh
matplotlib==3.6.3
    # via
    #   -r requirements/base.in
    #   pykin
mpmath==1.2.1
    # via sympy
msgpack==1.0.4
    # via trimesh
networkx==3.0
    # via
    #   pykin
    #   trimesh
numpy==1.24.1
    # via
    #   -r requirements/base.in
    #   contourpy
    #   gym
    #   mapbox-earcut
    #   matplotlib
    #   pykin
    #   python-fcl
    #   scipy
    #   shapely
    #   trimesh
packaging==23.0
    # via matplotlib
pillow==9.4.0
    # via
    #   matplotlib
    #   trimesh
pkgutil-resolve-name==1.3.10
    # via jsonschema
pybullet==3.2.5
    # via -r requirements/base.in
pycollada==0.7.2
    # via trimesh
pyglet==1.5.27
    # via trimesh
pykin==1.2.0
    # via -r requirements/base.in
pyparsing==3.0.9
    # via matplotlib
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pycollada
python-fcl==0.7.0.4
    # via pykin
pyyaml==6.0
    # via pykin
requests==2.28.2
    # via trimesh
rtree==1.0.1
    # via trimesh
scipy==1.10.0
    # via trimesh
shapely==2.0.1
    # via trimesh
six==1.16.0
    # via python-dateutil
svg-path==6.2
    # via trimesh
sympy==1.11.1
    # via trimesh
tqdm==4.64.1
    # via pykin
trimesh[easy]==3.18.1
    # via pykin
urllib3==1.26.14
    # via requests
xxhash==3.2.0
    # via trimesh
zipp==3.12.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
